Rewrite entire settings page using lessons learned
Removing the dependency on the settings page provided the ability to
build an overlay on any FL page to display the settings. The general
functionality is identical to what was built in the first version of
the page. However, now, FLES has a lot more control of how, and what
to present within the DIV without being locked into the constraints
of the FL settings page.