Add customization to view images inline in groups
The feature already exists in writing posts, and I was a bit annoyed
that it didn't exist in group posts (at least as a toggle) so that it
could be activated at will, and disabled on a page refresh. This limits
extraneous load on the server by having the feature enabled permanently.

When the feature is enabled, the user will see two new links on the
group discussion page in the right sidebar under 'Report Discussion'.
One is to enable viewing of images in the original post, and the other
is to view images in the thread. Once the user clicks on either, the
page is updated with the content.

On page reload or comment page change, the feature will have to be
reenabled.

Closes: https://github.com/unnaturaldevelopment/fles/issues/9