Add customization to quote in group discussion
The customization adds the ability to quote the original poster, or a
commenter in a group discussion. When the ability is enabled, the user
can select the text that they want to quote, copy the text either via
the keyboard (CTRL-C, or right click, and select copy). Once the text
is copied, the user clicks on 'Reply' when replying to the original
poster, or 'Multi-Reply' when replying to a commentor. The quoted text
should automatically populate in the message box, with the poster or
commentor's name prepended with the '@' to enable notification.

Closes: https://github.com/unnaturaldevelopment/fles/issues/14