Fix bug in Firefox on main group page
The multi-reply code was trying to run on the main group page, and
because it was unable to find the correct element, it was throwing
errors in the console. I added another partition of functions, and now
each of the partitions contain functions that are relevant to the
DOM of each.