Add feature to mention picture owner in comment
This is a continuation of the effort taken for group posts in providing
the user of FLES the ability to mention members easily when commenting.
I adjusted the multi-reply function so that I could take advantage of it
for this feature by adding a simply if/else for the comment box variant
in the picture page.

Closes: #23