Add feature to collapse list of user's writings
The user interface to view the list of a user's writings is horrible,
primarily because the index includes the full body of the writing text.
This feature compiles the entire list of writings, sorts them by date,
and removes the body text, leaving only the title and metadata for the
post.

A viewer can now discover more writing by a user, because all of the
content is available on one page.

Closes: #26